Former-commit-id: b9d95d79542a1eba5784207411a9f08260780059

+++ rec-common/src/main/java/es/capgemini/pfs/metrica/MetricaValidator.java
@@ -165,7 +165,7 @@ public class MetricaValidator {
                     continue;
                 }
 
-                if (nivel < 1 || nivel > 100) {
+                if (nivel.intValue() < 1 || nivel.intValue() > 100) {
                     return false;
                 }
             }
